GanHuo rename RealS(hit)
